Integração das funções de análise léxica, execução de expressões RPN, leitura, escrita de arquivos e exibição de resultados.
