+++ tokens_gerados.txt
@@ -1,10 +1,11 @@
-3.0 2.0 +
-10.0 4.0 -
-2.0 3.0 *
-9.0 2.0 /
-10.0 3.0 %
-2.0 3 ^
-1.0 2.0 + 3.0 4.0 * /
-5.0 MEM
+6.0 2.0 *
+12.0 4.0 /
+9.0 5.0 -
+3.0 3 ^
+15.0 4.0 %
+2.0 2.0 + 5.0 1.0 - *
+7.0 MEM
 MEM
-5 RES
+2 RES
+3.0 2.0 + 10.0 5.0 / *
+12.0 4.0 / 3 5 * +
